Attempt to fix with_defaults bug in ParameterFrame
Calling ParameterFrame(with_defaults=True) causes an error.

This commit fixes that by checking for an empty iterable of defaults
before attempting to apply them.

The new test works when run in isolation, however it fails when run as
part of the test suite. I believe this is because of the class method:
'set_default_parameters'. This sets defaults on the ParameterFrame
class. This is done in the Configuration class.

Hence, if the new test is run before importing from
'bluemira.base.config' it passes, if it's run after importing, it fails.