Add ComponentManager class (#1330)
* Add missing docstrings to parameter_frame module

* Add ComponentManager class

* Fix typing and params logic in Builder

* Use ComponentManager base in new builders

* Raise BuilderError on bad builder params input

* Remove unused import

* Restore BuildConfig type to base.builder

This should get the documentation building. We should revisit this type
in future, as it upsets MyPy.

* Add missing license header comment

* Remove tests with import errors

Removes tests that test code that has been removed as part of the
builder/designer refactor

* Add 'typeguard' to install_requires

It was already in requirements.txt, but had not been put in the setup.py

* Remove builder/shapes as no longer functional

* Remove test_design as Design class now removed

* Make Cryostat a ComponentManager

No extra attributes at the moment.