Add vacuum vessel to EUDEMO reactor example
This is currently not working because the vacuum vessel shape is not
valid!

After building the first section of the VV, we then rotate that section
360 around the center of the reactor. We then get a closed VV shape.
However, for every rotation we make of the section, the new section is
not valid. This seems to be an issue down to the FreeCAD level.

The fact that there are two disjointed shells in the section could be
part of the issue. Due to our Matryoshka geometry structure, when we
rotate we take the shape's boundary (a shell's boundary is a face, a
face's boundary is a wire, etc.). The face we get when taking the
shell's boundary is a fill between the two shells - which feels wrong
straight away. Taking the boundary of that gets us a (single!) wire,
which traces the inner and outer edges of the face! Seems like it does a
loop-the-loop. All this is in the xz-dimension.

FreeCAD does not like rotating that wire, and it returns an invalid
shape, which throws an error when we try to draw it.

A possible solution to this is to rotate the inner and outer face of the
VV separately, then FreeCAD may not get as confused about what the
boundaries of the shape should be.