Fix test_nlopt case failing due to mock
Because we were mocking out NLOpt, when we later attempted to retrieve
the bounds set on the mock, a MagicMock was returned. This caused an
error when attempting to use these bounds when calculating an
approximate derivative.

Fix by setting an analytical gradient, so we don't need to use an
approximate derivative.